use static_cast in PayloadHeader::mProtocolID initializer

+++ src/transport/raw/MessageHeader.h
@@ -583,7 +583,7 @@ private:
     Optional<uint16_t> mVendorId;
 
     /// Protocol identifier
-    uint16_t mProtocolID = Protocols::kProtocol_NotSpecified;
+    uint16_t mProtocolID = static_cast<uint16_t>(Protocols::kProtocol_NotSpecified);
 
     /// Bit flag indicators for CHIP Exchange header
     Header::ExFlags mExchangeFlags;
